mother slime + balance

+++ android/battle/src/main/java/com/game7th/battle/balance/SwipeBalance.kt
@@ -89,7 +89,7 @@ data class MotherSlimeBalance(
         val hp: PolinomBalance,
         val damage: PolinomBalance,
         val a2l: Int,
-        val regen: Int
+        val regen: Float
 )
 
 data class CitadelWarlockBalance(
